--- model/common/tests/metric_tests/utils.py
@@ -1,26 +1,0 @@
-# ICON4Py - ICON inspired code in Python and GT4Py
-#
-# Copyright (c) 2022, ETH Zurich and MeteoSwiss
-# All rights reserved.
-#
-# This file is free software: you can redistribute it and/or modify it under
-# the terms of the GNU General Public License as published by the
-# Free Software Foundation, either version 3 of the License, or any later
-# version. See the LICENSE.txt file at the top-level directory of this
-# distribution for a copy of the license or check <https://www.gnu.org/licenses/>.
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-import gt4py.next.program_processors.modular_executor
-
-
-def is_otf(backend) -> bool:
-    # want to exclude python backends:
-    #   - cannot run on embedded: because of slicing
-    #   - roundtrip is very slow on large grid
-    if hasattr(backend, "executor"):
-        if isinstance(
-            backend.executor, gt4py.next.program_processors.modular_executor.ModularExecutor
-        ):
-            return True
-    return False
